cos/tx_completion.rs: colocate 4 park-counter + timer-wheel tests with production (#984 P3 Phase 6a)
Append 4 tests to the existing cos/tx_completion.rs::tests block:
park_counter_queue_token_starvation_ticks_only_its_reason_on_exact,
park_counter_root_token_starvation_ticks_only_its_reason,
timer_wheel_cascades_long_parked_queue,
timer_wheel_wakes_short_parked_queue.

Adds select_cos_guarantee_batch + select_exact_cos_guarantee_queue_with_fast_path
imports (cos/queue_service) — the timer-wheel tests drive a service
pass to verify wakeup ordering after cascade.

Drops 5 stale cfg-test imports from tx/mod.rs (Codex round-1):
select_cos_guarantee_batch, select_exact_cos_guarantee_queue_with_fast_path,
advance_cos_timer_wheel, park_cos_queue, COS_TIMER_WHEEL_TICK_NS.

Rebased onto current master after #1012 merged.
cargo test --bins 865/0/2.
